refactor: clean up sandboxes components, remove unused files and re-think sandboxes router strategy
