+++ src/app/dashboard/@header/[teamIdOrSlug]/page.tsx
@@ -2,5 +2,3 @@
 export default function TeamDashboardHeaderContent() {
   return null
 }
-
-
