+++ src/app/dashboard/@header/[teamIdOrSlug]/sandboxes/page.tsx
@@ -1,16 +1,3 @@
-import { LiveSandboxCounterServer } from '@/features/dashboard/sandboxes/live-counter.server'
+import SandboxesHeaderInjectableDefault from './default'
 
-interface SandboxesHeaderInjectableCatchAllDefaultProps {
-  params: Promise<{ teamIdOrSlug: string }>
-}
-
-export default function SandboxesHeaderInjectableCatchAllDefault({
-  params,
-}: SandboxesHeaderInjectableCatchAllDefaultProps) {
-  return (
-    <LiveSandboxCounterServer
-      params={params}
-      className="top-5 absolute right-17 max-md:hidden"
-    />
-  )
-}
+export default SandboxesHeaderInjectableDefault
