+++ src/app/dashboard/[teamIdOrSlug]/sandboxes/@monitoring/page.tsx
