+++ src/app/dashboard/[teamIdOrSlug]/sandboxes/tabs.tsx
@@ -4,6 +4,7 @@ import { PROTECTED_URLS } from '@/configs/urls'
 import { cn } from '@/lib/utils'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/ui/primitives/tabs'
 import { ActivityIcon, LayoutListIcon } from 'lucide-react'
+import Link from 'next/link'
 import { useParams, usePathname, useSearchParams } from 'next/navigation'
 import { ReactNode } from 'react'
 
@@ -43,17 +44,6 @@ export default function SandboxesTabs({ children }: SandboxesTabsProps) {
   return (
     <Tabs
       value={activeTab?.value}
-      onValueChange={(value) => {
-        const newTab = TABS.find((tab) => tab.value === value)
-
-        if (newTab?.value === activeTab?.value) return
-
-        window.history.pushState(
-          null,
-          '',
-          `${PROTECTED_URLS.SANDBOXES(teamIdOrSlug, newTab?.value as 'list' | 'monitoring')}`
-        )
-      }}
       className="min-h-0 w-full flex-1 pt-3.5 h-full"
     >
       <TabsList className="bg-bg z-30 w-full justify-start pl-6">
@@ -63,9 +53,12 @@ export default function SandboxesTabs({ children }: SandboxesTabsProps) {
             layoutkey="tabs-indicator-sandboxes"
             value={tab.value}
             className="w-fit flex-none"
+            asChild
           >
-            <tab.icon className="size-3.5" />
-            {tab.value.charAt(0).toUpperCase() + tab.value.slice(1)}
+            <Link href={tab.url(teamIdOrSlug)} prefetch>
+              <tab.icon className="size-3.5" />
+              {tab.value.charAt(0).toUpperCase() + tab.value.slice(1)}
+            </Link>
           </TabsTrigger>
         ))}
       </TabsList>
